[core][lambda] split firehose into its own module
Break out the Firehose logic into its own class, this has several advantages:

* Clear deliniation between core rule processing logic and the Firehose processing
* Caching of disabled log types
* Segmented testing
* Switch config wording from disabled to enabled for firehose logs
