+++ tests/unit/stream_alert_rule_processor/test_handler.py
@@ -36,7 +36,6 @@ from tests.unit.stream_alert_rule_processor.test_helpers import (
 )
 
 
-@patch('stream_alert.rule_processor.handler.MAX_BACKOFF_ATTEMPTS', 1)
 class TestStreamAlert(object):
     """Test class for StreamAlert class"""
 
@@ -214,291 +213,6 @@ class TestStreamAlert(object):
 
         load_payload_mock.assert_called()
 
-    @patch('stream_alert.rule_processor.handler.LOGGER')
-    @mock_kinesis
-    def test_firehose_record_delivery(self, mock_logging):
-        """StreamAlert Class - Firehose Record Delivery"""
-        self.__sa_handler.firehose_client = boto3.client(
-            'firehose', region_name='us-east-1')
-
-        test_event = convert_events_to_kinesis([
-            # unit_test_simple_log
-            {
-                'unit_key_01': 1,
-                'unit_key_02': 'test'
-            },
-            {
-                'unit_key_01': 2,
-                'unit_key_02': 'test'
-            },
-            # test_log_type_json_nested
-            {
-                'date': 'January 01, 3005',
-                'unixtime': '32661446400',
-                'host': 'my-host.name.website.com',
-                'data': {
-                    'super': 'secret'
-                }
-            }
-        ])
-
-        delivery_stream_names = ['streamalert_data_test_log_type_json_nested',
-                                 'streamalert_data_unit_test_simple_log']
-
-        # Setup mock delivery streams
-        for delivery_stream in delivery_stream_names:
-            self.__sa_handler.firehose_client.create_delivery_stream(
-                DeliveryStreamName=delivery_stream,
-                S3DestinationConfiguration={
-                    'RoleARN': 'arn:aws:iam::123456789012:role/firehose_delivery_role',
-                    'BucketARN': 'arn:aws:s3:::kinesis-test',
-                    'Prefix': '{}/'.format(delivery_stream),
-                    'BufferingHints': {
-                        'SizeInMBs': 123,
-                        'IntervalInSeconds': 124
-                    },
-                    'CompressionFormat': 'Snappy',
-                }
-            )
-
-        with patch.object(self.__sa_handler.firehose_client, 'put_record_batch') as firehose_mock:
-            firehose_mock.return_value = {'FailedPutCount': 0}
-            self.__sa_handler.run(test_event)
-
-            firehose_mock.assert_called()
-            assert_true(mock_logging.info.called)
-
-    @patch('stream_alert.rule_processor.handler.LOGGER')
-    @mock_kinesis
-    def test_firehose_record_delivery_failed_put_count(self, mock_logging):
-        """StreamAlert Class - Firehose Record Delivery - Failed Put Count"""
-        self.__sa_handler.firehose_client = boto3.client(
-            'firehose', region_name='us-east-1')
-
-        test_event = convert_events_to_kinesis([
-            # unit_test_simple_log
-            {
-                'unit_key_01': 1,
-                'unit_key_02': 'test'
-            },
-            {
-                'unit_key_01': 2,
-                'unit_key_02': 'test'
-            },
-            # test_log_type_json_nested
-            {
-                'date': 'January 01, 3005',
-                'unixtime': '32661446400',
-                'host': 'my-host.name.website.com',
-                'data': {
-                    'super': 'secret'
-                }
-            }
-        ])
-
-        delivery_stream_names = ['streamalert_data_test_log_type_json_nested',
-                                 'streamalert_data_unit_test_simple_log']
-
-        # Setup mock delivery streams
-        for delivery_stream in delivery_stream_names:
-            self.__sa_handler.firehose_client.create_delivery_stream(
-                DeliveryStreamName=delivery_stream,
-                S3DestinationConfiguration={
-                    'RoleARN': 'arn:aws:iam::123456789012:role/firehose_delivery_role',
-                    'BucketARN': 'arn:aws:s3:::kinesis-test',
-                    'Prefix': '{}/'.format(delivery_stream),
-                    'BufferingHints': {
-                        'SizeInMBs': 123,
-                        'IntervalInSeconds': 124
-                    },
-                    'CompressionFormat': 'Snappy',
-                }
-            )
-
-        with patch.object(self.__sa_handler.firehose_client, 'put_record_batch') as firehose_mock:
-            firehose_mock.side_effect = [
-                {
-                    'FailedPutCount': 3,
-                    'RequestResponses': [
-                        {
-                            "ErrorCode": "ServiceUnavailableException",
-                            "ErrorMessage": "Slow down."
-                        },
-                        {
-                            "ErrorCode": "ServiceUnavailableException",
-                            "ErrorMessage": "Slow down."
-                        },
-                        {
-                            "ErrorCode": "ServiceUnavailableException",
-                            "ErrorMessage": "Slow down."
-                        }
-                    ]
-                },
-                {
-                    'FailedPutCount': 3,
-                    'RequestResponses': [
-                        {
-                            "ErrorCode": "ServiceUnavailableException",
-                            "ErrorMessage": "Slow down."
-                        },
-                        {
-                            "ErrorCode": "ServiceUnavailableException",
-                            "ErrorMessage": "Slow down."
-                        },
-                        {
-                            "ErrorCode": "ServiceUnavailableException",
-                            "ErrorMessage": "Slow down."
-                        }
-                    ]
-                },
-                {
-                    'FailedPutCount': 0,
-                    'RequestResponses': [
-                        {
-                            "RecordId": "12345678910",
-                            "ErrorCode": "None",
-                            "ErrorMessage": "None"
-                        },
-                        {
-                            "RecordId": "12345678910",
-                            "ErrorCode": "None",
-                            "ErrorMessage": "None"
-                        },
-                        {
-                            "RecordId": "12345678910",
-                            "ErrorCode": "None",
-                            "ErrorMessage": "None"
-                        }
-                    ]
-                }]
-            self.__sa_handler.run(test_event)
-
-            firehose_mock.assert_called()
-            assert_true(mock_logging.info.called)
-
-    @patch('stream_alert.rule_processor.handler.LOGGER')
-    def test_firehose_limit_record_size(self, mock_logging):
-        """StreamAlert Class - Firehose - Record Size Check"""
-        test_events = [
-            # unit_test_simple_log
-            {
-                'unit_key_01': 1,
-                'unit_key_02': 'test' * 250001  # is 4 bytes higher than max
-            },
-            {
-                'unit_key_01': 2,
-                'unit_key_02': 'test'
-            },
-            # test_log_type_json_nested
-            {
-                'date': 'January 01, 3005',
-                'unixtime': '32661446400',
-                'host': 'my-host.name.website.com',
-                'data': {
-                    'super': 'secret'
-                }
-            }
-        ]
-
-        self.__sa_handler._limit_record_size(test_events)
-
-        assert_true(len(test_events), 2)
-        assert_true(mock_logging.error.called)
-
-    @patch('stream_alert.rule_processor.handler.LOGGER')
-    @mock_kinesis
-    def test_firehose_record_delivery_failure(self, mock_logging):
-        """StreamAlert Class - Firehose Record Delivery - Failed PutRecord"""
-        class MockFirehoseClient(object):
-            @staticmethod
-            def put_record_batch(**kwargs):
-                return {
-                    'FailedPutCount': len(kwargs.get('Records')),
-                    'RequestResponses': [
-                        {
-                            'RecordId': '12345',
-                            'ErrorCode': '300',
-                            'ErrorMessage': 'Bad message!!!'
-                        },
-                    ]
-                }
-
-        self.__sa_handler.firehose_client = MockFirehoseClient()
-
-        test_event = convert_events_to_kinesis([
-            # unit_test_simple_log
-            {
-                'unit_key_01': 1,
-                'unit_key_02': 'test'
-            },
-            {
-                'unit_key_01': 2,
-                'unit_key_02': 'test'
-            },
-            # test_log_type_json_nested
-            {
-                'date': 'January 01, 3005',
-                'unixtime': '32661446400',
-                'host': 'my-host.name.website.com',
-                'data': {
-                    'super': 'secret'
-                }
-            }
-        ])
-
-        self.__sa_handler.run(test_event)
-        assert_true(mock_logging.error.called)
-
-    def test_firehose_sanitize_keys(self):
-        """StreamAlert Class - Firehose - Sanitize Keys"""
-        # test_log_type_json_nested
-        test_event = {
-            'date': 'January 01, 3005',
-            'unixtime': '32661446400',
-            'host': 'my-host.name.website.com',
-            'data': {
-                'super-duper': 'secret',
-                'sanitize_me': 1,
-                'example-key': 1,
-                'moar**data': 2,
-                'even.more': 3
-            }
-        }
-
-        expected_sanitized_event = {
-            'date': 'January 01, 3005',
-            'unixtime': '32661446400',
-            'host': 'my-host.name.website.com',
-            'data': {
-                'super_duper': 'secret',
-                'sanitize_me': 1,
-                'example_key': 1,
-                'moar__data': 2,
-                'even_more': 3
-            }
-        }
-
-        sanitized_event = self.__sa_handler.sanitize_keys(test_event)
-        assert_equal(sanitized_event, expected_sanitized_event)
-
-    def test_firehose_segment_records_by_size(self):
-        """StreamAlert Class - Firehose - Segment Large Records"""
-
-        record_batch = [
-            # unit_test_simple_log
-            {'unit_key_01': 2, 'unit_key_02': 'testtest' * 10000}
-            for _
-            in range(100)]
-
-        sized_batches = []
-
-        for sized_batch in self.__sa_handler._segment_records_by_size(record_batch):
-            sized_batches.append(sized_batch)
-
-        assert_true(len(str(sized_batches[0])) < 4000000)
-        assert_equal(len(sized_batches), 4)
-        assert_true(isinstance(sized_batches[3][0], dict))
-
     @mock_kinesis
     def test_firehose_record_delivery_disabled_logs(self):
         """StreamAlert Class - Firehose Record Delivery - Disabled Logs"""
@@ -537,24 +251,3 @@ class TestStreamAlert(object):
             self.__sa_handler.run(test_event)
 
             firehose_mock.assert_not_called()
-
-    @patch('stream_alert.rule_processor.handler.LOGGER')
-    @mock_kinesis
-    def test_firehose_record_delivery_client_error(self, mock_logging):
-        """StreamAlert Class - Firehose Record Delivery - Client Error"""
-        self.__sa_handler.firehose_client = boto3.client(
-            'firehose', region_name='us-east-1')
-
-        test_events = [
-            # unit_test_simple_log
-            {'unit_key_01': 2, 'unit_key_02': 'testtest'}
-            for _
-            in range(10)]
-
-        self.__sa_handler._firehose_request_helper('invalid_stream',
-                                                   test_events)
-
-        missing_stream_message = 'Client Error ... An error occurred ' \
-            '(ResourceNotFoundException) when calling the PutRecordBatch ' \
-            'operation: Stream invalid_stream under account 123456789012 not found.'
-        assert_true(mock_logging.error.called_with(missing_stream_message))
